Move package types to own directory
