+++ src/packagetype/class-theme.php
@@ -1,20 +1,22 @@
 <?php
 /**
- * Package_Theme class
+ * Theme class
  *
  * @package SatisPress
  * @license GPL-2.0-or-later
  * @since 0.2.0
  */
 
-namespace SatisPress;
+namespace SatisPress\PackageType;
+
+use SatisPress\Package;
 
 /**
  * Theme package class.
  *
  * @since 0.2.0
  */
-class Package_Theme extends Package {
+class Theme extends Package {
 	/**
 	 * Base path where packages are cached.
 	 *
